% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggconsulting-package.R
\docType{package}
\name{ggconsulting-package}
\alias{ggconsulting}
\alias{ggconsulting-package}
\title{ggconsulting: Executive-Grade Themes and Helpers for 'ggplot2'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An opinionated 'ggplot2' extension for executive-grade consulting output. Ships archetype themes, palettes, scale helpers, locale-aware label helpers, and a data-aware polish layer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/viniciusoike/ggconsulting}
  \item \url{https://viniciusoike.github.io/ggconsulting/}
  \item Report bugs at \url{https://github.com/viniciusoike/ggconsulting/issues}
}

}
\author{
\strong{Maintainer}: Vinicius Reginatto \email{viniciusoike@gmail.com}

}
\keyword{internal}
